ResourceManager 추가 ResourceManager는 각 오브젝트가 Texture를 가지고 있는것은 메모리 낭비가 되므로 이를 한꺼번에 관리해줄 Manager를 만듬

+++ WinApi2D/CPathManager.cpp
@@ -18,7 +18,7 @@ void CPathManager::init()
 	// 상위 폴더로
 	// + bin\\content\\
 
-	int iLen = wcslen(m_strContentPath);		// wchar_t 문자열의 길이 확인
+	int iLen = (int)wcslen(m_strContentPath);		// wchar_t 문자열의 길이 확인
 
 	for (int i = iLen - 1; i >= 0; i--)
 	{
